git: Create blank .gitconfig
Using the XDG Base Directory Specification might be a good idea, but I'm
not quite bothered enough by the number of dotfiles in $HOME to use that
instead of the traditional vomit-dotfiles-all-over-$HOME approach.

There's also the fact that the XDG spec does not appear to be widely
supported enough to totally eliminate dotfiles in $HOME. For example,
vim and bash are still hardcoded to read from files/directories in
$HOME, and setting up the right environment variables to get them to
work with the XDG spec might be more trouble than it's worth at this
time.