vim: Add empty vimrc
~/.vim/vimrc is used instead of ~/.vimrc so all vim-related files can
live in one place (except for neovim's ~/.config/nvim/ symlink, but
that's just a symlink).

Vim learned to read ~/.vim/vimrc/ in version 7.3.1178 (commit
22971aaa4530d6e8cb2c9c0c19b4f291d8442d61), committed on 2013-06-12. That
is hopefully old enough that all systems I can be reasonably expected to
work on will include that patch.

Systems that include a default ~/.vimrc might cause momentary confusion
since that will get read instead of the vimrc in this repo, but that is
an easy fix that can also be taken care of by a setup script if/when I
get around to writing one.