*vim: Set up vim-plug
Several plugin managers are available, but vim-plug seems to have a good
set of features:

  - No rtp/nocompatible[0]/filetype[1]/etc. boilerplate
  - Parallel/async installs/updates if available
  - Fast
  - Clean syntax
  - Easier to use than the built-in packaging feature
  - Other things I'm probably forgetting/haven't learned about
  - I'm lazy and already somewhat familiar with it.

Adding something to autoinstall vim-plug if it isn't present could be
nice, but that situation probably doesn't happen frequently enough (and
isn't annoying enough) to be worth bothering with. There will probably
need to be a setup script for the dotfiles repo anyways, since it's
(currently) intended to be used as a bare repo, so downloading vim-plug
could easily fit in there without cluttering the vimrc.

[0]: It appears that Vim automatically sets "nocompatible" as soon as a
     user vimrc is found. This has been the case since at least Vim
     v7.0001, committed 2004-06-13.

[1]: "filetype plugin indent on" and "syntax enable" are taken care of
     automatically by plug#end().