*vim: Enable persistent undo if available
A mkdir() call could have been added to ensure that ~/.vim/undodir
exists, but the assumption is that that is something that should only
need to happen once, and so is better handled by a setup script of some
kind. In addition, this also assumes that there is little or no reason
for the undo directory to be deleted or otherwise rendered inaccessible.

Using a mkdir() call has a few issues to work around, too. Up until Vim
8.0.1708 (and some early-ish version of Neovim), attempting to makedir()
an existing directory would result in an error, even with the 'p'
argument. Unfortunately, I'm not confident that I can rely on Vim 8+
being installed everywhere my dotfiles are going to be used.

isdirectory() could be used to only call mkdir() if the directory isn't
already present, but that is subject to race conditions. The chance of
actually hitting the race semi-consistently is probably quite low, but
the solution isn't exactly elegant.

silent! could be used to hide any errors from mkdir(), but that could
mean that errors not related to being unable to create the directory
could be hidden (if such errors exist).