*vim: Set undo limit to i32::MAX
Effectively set an unlimited undo history. This assumes that it would be
extremely difficult, if not impossible, to get an undo history large
enough for memory/disk consumption and/or startup time to be an issue,
especially since file writes from non-vim sources (e.g., git checkouts)
appear to wipe the undo history. If the history length does become an
issue, it hopefully will be fairly easy to delete the offending undo
file or otherwise truncate the history.

Neovim v0.4.3 gives an error if a value larger than i32::MAX is passed.
Vim 8.0.0218 and later set undolevels to min(i64::MAX, <input value>)
(or probably i32::MAX for 32-bit builds), while earlier versions seem to
either let undolevels overflow or set it to 0.