*vim: Show absolute line numbers
Helpful for getting a general sense of where you are in a file,
especially when trying to communicate locations to other people.

The line number is also shown in the ruler, but that's slightly out of
the way compared to just looking left, and the single line number is a
bit less helpful for communicating ranges of line numbers.