git: Show diff in commit message template
Can be helpful for a quick reminder of exactly what changes were made in
a commit without having to switch windows and/or run git diff/etc.

Choosing to show both staged and unstaged changes is more or less an
arbitrary choice. Showing staged changes at least is probably a good
idea, but I'm hoping showing unstaged changes won't be a big problem.

A potential downside is that it can make committing larger diffs slow,
but hopefully those will be rare enough that it won't pose a huge
problem.