*vim: Jump to last known cursor position on file open
Taken from defaults.vim, which is included with a vim install as far as
I know.

The comment explaining the code is:

    When editing a file, always jump to the last known cursor position.
    Don't do it when the position is invalid, when inside an event handler
    (happens when dropping a file on gvim) and for a commit message (it's
    likely a different one than last time).

A possible improvement might be to also disable this behavior for diff
hunks (e.g., to avoid jumping to nonsensical places when editing git add
-p hunks). This makes some operations easier (i.e., the aforementioned
git add -p hunk editing), but can mess with others (editing normal
diffs). The former happens more frequently than the latter for my
current mix of work, but loading to an unpredictable part of the file
isn't annoying enough (yet) to make the change.